+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.14] - 2024-12-19
+
+### Changed
+- People list: People are now sorted alphabetically by last name (with first name as secondary sort)
+- People list: Company name is now displayed below each person's name (shows current company or most recent)
+
 ## [1.0.13] - 2024-12-19
 
 ### Added